added CreditController, exception handling for delete movie

+++ src/main/java/com/bmdb/web/MovieController.java
@@ -3,7 +3,10 @@ package com.bmdb.web;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.bmdb.business.Movie;
 import com.bmdb.db.MovieRepo;
@@ -37,8 +40,29 @@ public class MovieController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable int id) {
-		movieRepo.deleteById(id);
+	public Optional<Movie> delete(@PathVariable int id) {
+		Optional<Movie> movie = movieRepo.findById(id);
+		if (movie.isPresent()) {
+			try {
+				movieRepo.deleteById(id);
+			}
+			catch (DataIntegrityViolationException dive) {
+				// catch dive when movie exists as fk on another table
+				System.err.println(dive.getRootCause().getMessage());
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+						"Foreign Key Constraint Issue - movie id: "+id+ " is "
+								+ "referred to elsewhere.");
+			}
+			catch (Exception e) {
+				e.printStackTrace();
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, 
+						"Exception caught during movie delete.");
+			}
+		}
+		else {
+			System.err.println("Movie delete error - no movie found for id:"+id);
+		}
+		return movie;
 	}
 
 }
